offset 可以为负数

+++ nanos-lite/include/fs.h
@@ -1,6 +1,7 @@
 #ifndef __FS_H__
 #define __FS_H__
 
+#include <stdio.h>
 #include <common.h>
 
 #ifndef SEEK_SET
@@ -12,7 +13,6 @@ size_t fs_write(int fd, const void *buf, size_t len);
 size_t fs_read(int fd, void *buf, size_t len);
 int fs_close(int fd);
 
-size_t fs_lseek(int fd, size_t offset, int whence);
-
+off_t fs_lseek(int fd, off_t offset, int whence);
 
 #endif
